% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_structures.R
\name{build_data_structures}
\alias{build_data_structures}
\title{Build the classic cointegration data structures, S and Z}
\usage{
build_data_structures(spec, data)
}
\arguments{
\item{spec}{rsci_spec, an rsci specification object.}

\item{data}{xts, a matrix with the data that should be fitted.}
}
\value{
a list with a list of S matrices and a list of Z matrices.
}
\description{
Build the classic cointegration data structures, S and Z
}
