% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_covariance.R
\name{calc_error_covariance}
\alias{calc_error_covariance}
\title{Calculate error covariances given data and other parameters}
\usage{
calc_error_covariance(alpha, beta, gamma, lambda, H_ec, h_ec)
}
\arguments{
\item{alpha}{tibble, a table with regime names and matrices}

\item{beta}{matrix, the cointegration paramters}

\item{gamma}{tibble, a table with regime names and matrices}

\item{lambda}{vector, the parameter vector that goes into the}
}
\value{
a list with covariance matrices for each regime
}
\description{
Calculate error covariances given data and other parameters
}
