% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_short_run_pars.R
\name{calc_short_run_pars}
\alias{calc_short_run_pars}
\title{Calculate short run parameters given data and other parameters}
\usage{
calc_short_run_pars(beta, omega, spec, data, regime_prob, H_sr, h_sr)
}
\arguments{
\item{beta}{matrix, the cointegration relations}

\item{omega}{list, a list with the covariance matrices for each regime}

\item{spec}{rsci_spec, a specification object.}

\item{data}{xts, the data the model should fit to.}

\item{lambda}{vector, parameters for the cointegration relations}
}
\value{
a list with the short run parameters, alpha and gamma for each regime
}
\description{
Calculate short run parameters given data and other parameters
}
