% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regime_residuals.R
\name{calculate_regime_residuals}
\alias{calculate_regime_residuals}
\title{Cacluate regime specific residuals}
\usage{
calculate_regime_residuals(Phi, beta, Z)
}
\arguments{
\item{Phi}{list, a list of least squares paramters for each regime}

\item{beta}{matrix, a matrix with the cointegration relations}

\item{Z}{list, a list of data matrices constructed with the build function.}
}
\value{
a list of matrices with regime residuals
}
\description{
Cacluate regime specific residuals
}
