% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_null_regime.R
\name{check_for_null_regime}
\alias{check_for_null_regime}
\title{Check for the number}
\usage{
check_for_null_regime(regprobs)
}
\arguments{
\item{regprobs}{list, a list of the xts regime probabilities}
}
\value{
Boolean TRUE if a regime is deamed neglected
}
\description{
Check for the number
}
