% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_closed_form_pars.R
\name{estimate_closed_form_pars}
\alias{estimate_closed_form_pars}
\title{Estimate model parameters}
\usage{
estimate_closed_form_pars(pars, model, ds, regprobs)
}
\arguments{
\item{pars}{vector, a vector of non linear parameters}

\item{model}{rsci_model object.}

\item{ds}{rsci_data object.}

\item{regprobs}{list, a list with regime probabilities.}
}
\value{
model
}
\description{
Estimate model parameters
}
