% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_error_covariances.R
\name{estimate_error_covariances}
\alias{estimate_error_covariances}
\title{Estimate the covariance matrices given filtered probabilities and
       other model parameters.}
\usage{
estimate_error_covariances(Phi, beta, Z, regprobs, H, h)
}
\arguments{
\item{Phi}{list, a list with the regime specific least squares matrices}

\item{beta}{matrix, a matrix with the cointegration vectors}

\item{Z}{list, a list with the data matrices}

\item{regprobs}{list, a list with vectors regime probabilities}

\item{H}{matrix, the restriction matrix}
}
\value{
the vectorized version of Omega
}
\description{
Estimate the covariance matrices given filtered probabilities and
       other model parameters.
}
