% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_least_squares.R
\name{estimate_least_squares}
\alias{estimate_least_squares}
\title{Estimate least squares parameters alpha and Gamma}
\usage{
estimate_least_squares(Z, beta, Omega, regprobs, H, h)
}
\arguments{
\item{Z}{list, the standard matrices Z0, Z1 and Z2}

\item{beta}{matrix, the matrix of cointegration relations}

\item{Omega}{list, a list of matrices corresponding to the error covariance
matrices by regime}

\item{regprobs}{list, a list with vectors of regime probabilities}

\item{H}{matrix, a matrix with the linear restrictions}

\item{h}{vector, the normalizing vector for linear restrictions}
}
\value{
the vector Phi with estimated least squares parameters.
}
\description{
Estimate least squares parameters alpha and Gamma
}
