% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_U.R
\name{get_U}
\alias{get_U}
\title{Construct U}
\usage{
get_U(Z, beta)
}
\arguments{
\item{Z}{list, the list with Z0, Z1 and Z2}

\item{beta}{matrix, the cointegration relations}
}
\value{
matrix with U
}
\description{
Construct U
}
