% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_Omega.R
\name{initial_Omega}
\alias{initial_Omega}
\title{Estimate initial parameters for Phi and Omega.}
\usage{
initial_Omega(nreg, Z, beta, regprobs, H, h, Phi = NULL)
}
\arguments{
\item{nreg}{integer, the number of regimes}

\item{Z}{list, the standard data structures Z0, Z1 and Z2}

\item{beta}{vector, a vector of probability paramters}

\item{regprobs}{list, a list with an xts time series of regime probabilities
for each regime.}

\item{H}{matrix, the selection matric for linear restrictions}

\item{h}{vector, the normalizing vector for linear restrictions}

\item{Phi}{list, a list with initial values for Phi.}
}
\value{
a list with initial regime Omega estimates.
}
\description{
Estimate initial parameters for Phi and Omega.
}
