% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_loglik.R
\name{int_loglik}
\alias{int_loglik}
\title{Internal calculation of the loglikelihood.}
\usage{
int_loglik(pars, Z, regprobs)
}
\arguments{
\item{pars}{an rsci_pars object}

\item{Z}{list, a list holding the data structures Z0, Z1, Z2}

\item{regprobs}{vector, the vedtor of regime probabilities}
}
\value{
a likelihood number
}
\description{
a logliklihood function that works internally and avoids
             rebuilding data structions at each evaluation.
}
