% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_build_linres_Omega.R
\name{rsci_build_linres_Omega}
\alias{rsci_build_linres_Omega}
\title{Construct linear restriction matrices for a set of common
       specifications.}
\usage{
rsci_build_linres_Omega(dim, nreg, regimes = NULL, name = "none")
}
\arguments{
\item{dim}{integer, dimension of the system}

\item{nreg}{integer, the number of regimes}

\item{regimes}{tibble, a table with the regime names and groups. Only needed
if the option "fixed_group" is chosen.}

\item{name}{string, name of the specification.}
}
\value{
a list with the restriction matrices H_ec and h_ec for the chosen
        standard specification.
}
\description{
Construct linear restriction matrices for a set of common
       specifications.
}
