% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_build_linres_Phi.R
\name{rsci_build_linres_Phi}
\alias{rsci_build_linres_Phi}
\title{Construct linear restriction matrices for a set of common
       specifications for alpha,}
\usage{
rsci_build_linres_Phi(dim, rank, lags, nreg, regimes = NULL,
  restriction = "none")
}
\arguments{
\item{dim}{integer, dimension of the system}

\item{rank}{integer, the number of cointegration relations}

\item{lags}{integer, the number of lags in the model}

\item{nreg}{integer, the number of regimes}

\item{regimes}{tibble, a table with the regime names and groups. Only needed
if fixed_group is used.}

\item{name}{string, name of the specification.
Options are}
}
\value{
a list with the restriction matrices Ha and ha
}
\description{
Construct linear restriction matrices for a set of common
       specifications for alpha,
}
