% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_data.R
\name{rsci_data}
\alias{rsci_data}
\title{Build the classic cointegration data structures Z0, Z1 and Z2.}
\usage{
rsci_data(rank, lags, data)
}
\arguments{
\item{rank}{integer, the cointegration rank}

\item{lags}{integer, the number of lags in the model}

\item{data}{xts, a matrix with the data that should be fitted.}
}
\value{
a list with a list of S matrices and a list of Z matrices.
}
\description{
Build the classic cointegration data structures Z0, Z1 and Z2.
}
