% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_fit.R
\name{rsci_fit}
\alias{rsci_fit}
\title{Fit a regime switching cointegration vector error correction model to
       data.}
\usage{
rsci_fit(spec, init_lambda, init_b, fn_b2beta, fn_regprobs, data,
  restrict_alpha = "none", restrict_gamma = "none",
  restrict_omega = "none")
}
\arguments{
\item{spec}{rsci_spec, a specification object}

\item{init_lambda}{vector a vector of}

\item{init_b}{vector, a vector of initial values for the estimated parts
of beta.}

\item{fn_b2beta}{function, a function that takes the values from b and
produces a matrix of size (p x r) coresponding to the long run
cointegration vectors.}

\item{data}{xts, a matrix with the data.}

\item{restrict_alpha}{string, a string with standard options for restricting
alpha to be the same across certain regimes.}

\item{fn_regprob}{function, a function that takes two arguments
\code{ fn_regprob(model, data) } where model is a rsci_model
object and data is an xts object with the time series data used in
rsci_fit. The function must return a simple vector of regime
probabilities that has length corresponding to the number of columns
in the regimes tibble in the rsci_spec object.}
}
\value{
an rsci_fit object
}
\description{
Fit a regime switching cointegration vector error correction model to
       data.
}
