% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_loglik.R
\name{rsci_loglik}
\alias{rsci_loglik}
\title{Calculate the logliklihood value some data given a model.}
\usage{
rsci_loglik(model, regprobs, data_struct)
}
\arguments{
\item{model}{rsci_model, a model object with the model in question}

\item{regprobs}{list, a list with regimes probabilities}

\item{data}{rsci_data, the data to be fitted}
}
\value{
the log-likelihood value
}
\description{
Calculate the logliklihood value some data given a model.
}
