% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_monte_carlo.R
\name{rsci_monte_carlo}
\alias{rsci_monte_carlo}
\title{Do a monte carlo simulation of the estimation procedure.}
\usage{
rsci_monte_carlo(model, sample_size, replications)
}
\arguments{
\item{model}{rsci_model, the model that we wish to simulate from and
subsequently estimate parameters on.}

\item{sample_size}{integer, the number of observations to simulate.}

\item{replications}{integer, the number of replications of the experiment.}

\item{burn}{integer, a number of periods to burn from the beginning of
the sample to avoid dependence on initial values.}
}
\value{
an rsci_mc object with all information on the experiment.
}
\description{
Do a monte carlo simulation of the estimation procedure.
}
