% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_spec.R
\name{rsci_spec}
\alias{rsci_spec}
\title{Create a regime switching cointegrated vector error correction model
       specification object.}
\usage{
rsci_spec(lags, rank, regimes)
}
\arguments{
\item{lags}{integer, the number of lags in levels in the model.}

\item{rank}{integer, the cointegration rank.}

\item{regimes}{tibble, a table with regime names and groups.}
}
\value{
an rsci_spec object holding the model specification without parameter
        values.
}
\description{
Create a regime switching cointegrated vector error correction model
       specification object.
}
