% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_lin_res_alpha.R
\name{std_lin_res_alpha}
\alias{std_lin_res_alpha}
\title{Construct linear restriction matrices for a set of common
       specifications for alpha}
\usage{
std_lin_res_alpha(dim, lags, restriction = "fixed")
}
\arguments{
\item{dim}{integer, dimension of the system}

\item{name}{string, name of the specification.
Options are
\itemize{
 \item{"fixed_gamma"}{Fix short run parameters across regimes}
 \item{"fixed_gamma_group"}{Fix short run parameters across regimes
                            belonging to a regime group.}
 \item{"fixed_alpha"}{Fix error correction parameters across regimes}
 \item{"fixed_alpha_group}{Fix error correction parameters across
                           regimes belonging to a regime group.}
 \item{"fixed"}{Fix all least squares parameters across regimes.}
 \item{"fixed_group"}{Fix all least squares parameters across groups.}
 \item{"none"}{The value of gamma changes with each regime}
}}
}
\value{
a list with the restriction matrices H_sr and h_sr
}
\description{
Construct linear restriction matrices for a set of common
       specifications for alpha
}
