% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_lin_res_omega.R
\name{std_lin_res_omega}
\alias{std_lin_res_omega}
\title{Construct linear restriction matrices for a set of common
       specifications}
\usage{
std_lin_res_omega(dim, lags, name)
}
\arguments{
\item{dim}{integer, dimension of the system}

\item{lags}{integer, the number of lags in the model}

\item{name}{string, name of the specification.
Options are
\itemize{
 \item{"fixed"}{Fix short run parameters across regimes}
 \item{"fixed_group"}{Fix short run parameters across regimes
                            belonging to a regime group.}
 \item{"fixed"}{Fix error correction parameters across regimes}
 \item{"fixed_group}{Fix error correction parameters across
                           regimes belonging to a regime group.}
 \item{"none"}{The value of lambda changes with each regime.}
}}
}
\value{
a list with the restriction matrices H_ec and h_ec for the chosen
        standard specification.
}
\description{
Construct linear restriction matrices for a set of common
       specifications
}
