% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unvech.R
\name{unvech}
\alias{unvech}
\title{Unvech undoes the vech operation}
\usage{
unvech(vec)
}
\arguments{
\item{vec}{vector, lower diagonal elements of a symmetric matrix.}
}
\value{
a symmetric matrix with from the vech vector
}
\description{
Unvech undoes the vech operation
}
