% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_Omega_2_list_Omega.R
\name{vec_Omega_2_list_Omega}
\alias{vec_Omega_2_list_Omega}
\title{Transform the vectorized version of Omega into the list version}
\usage{
vec_Omega_2_list_Omega(vecOmega, dim, regs)
}
\arguments{
\item{vecOmega}{vector, the vectorized version of Omega}

\item{dim}{integer, the dimension of the system}

\item{regs}{integer, the number of regimes}
}
\value{
the list Omegas by regime
}
\description{
Transform the vectorized version of Omega into the list version
}
