% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_Phi_2_list_Phi.R
\name{vec_Phi_2_List_Phi}
\alias{vec_Phi_2_List_Phi}
\title{Trandsform the vector version of Phi to the list version}
\usage{
vec_Phi_2_List_Phi(vecPhi, dim, rank, lags, nreg)
}
\arguments{
\item{vecPhi}{vector, the vectorized version of Phi}

\item{dim}{integer, the dimension of the system}

\item{rank}{integer, the cointegration rank}

\item{lags}{integer, the number of lags in the model}

\item{nreg}{integer, the number of regimes}
}
\value{
the list of Phi by regime
}
\description{
Trandsform the vector version of Phi to the list version
}
