% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vech.R
\name{vech}
\alias{vech}
\title{Vech operator as in Magnus and Neudecker.}
\usage{
vech(mat)
}
\arguments{
\item{mat}{matrix.}
}
\value{
a vector with the lower triangle elements of the input matrix,
        including the diagonal.
}
\description{
Vech operator as in Magnus and Neudecker.
}
