% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_fit.R
\name{rsci_fit}
\alias{rsci_fit}
\title{Fit a regime switching cointegration vector error correction model to
       data.}
\usage{
rsci_fit(model, data, init_pars, fn_update, ...)
}
\arguments{
\item{model}{rsci_model object, the object containing the initial model.}

\item{init_pars}{vector or number, initial parameters for the freely varying
elements in beta and the parameters in the probability function.}

\item{fn_update}{function, a function that takes the values from a vector
pars and allocates the parameters correctly to the model object.
Returns a model object. The function must take the following arguments
fn_update(pars, model). These are the  free parameters in beta and
the probability functions.}

\item{...}{additional information to be passed to optim().}
}
\value{
an rsci_fit object
}
\description{
This function takes data
}
