% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_model.R
\name{rsci_model}
\alias{rsci_model}
\title{Create a regime switching cointegrated vector error correction}
\usage{
rsci_model(spec, parameters, fn_regprob)
}
\arguments{
\item{spec}{rsci_spec, an rsci_spec object with the model specification}

\item{fn_regprob}{function, a function that takes two arguments
\code{ fn_regprob(pars, data) } where pars is a rsci_pars
object and data is an xts object with the time series data used in
rsci_fit. The function must return a simple vector of regime
probabilities that has length corresponding to the number of columns
in the regimes tibble in the rsci_spec object.}
}
\value{
an rsci_model object holding the model
}
\description{
Create a regime switching cointegrated vector error correction
}
