% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsci_simulate.R
\name{rsci_simulate}
\alias{rsci_simulate}
\title{Simulate a data series from a regime switching error correction model}
\usage{
rsci_simulate(model, sample_size, burn = 0)
}
\arguments{
\item{model}{rsci_model obj}

\item{sample_size}{integer, the size of the sample.}

\item{burn}{integer, a number of periods to burn from the begining of the
sample}
}
\value{
a matrix with simulated values for the model
}
\description{
Simulate a data series from a regime switching error correction model
}
